# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List all automations
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create a new automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automations/{id}/run:
    post:
      operationId: runAutomation
      tags: [automations]
      summary: Trigger an automation to run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRunResult"
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    put:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHabitBody"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/{id}/complete:
    post:
      operationId: completeHabit
      tags: [habits]
      summary: Mark a habit as completed for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated habit with streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
...
components:
  schemas:
...
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        script:
          type: string
        trigger:
          type: string
          enum: [manual, scheduled, event]
        status:
          type: string
          enum: [active, inactive]
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        lastResult:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - script
        - trigger
        - status
        - createdAt
        - updatedAt
    CreateAutomationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        script:
          type: string
        trigger:
          type: string
          enum: [manual, scheduled, event]
        status:
          type: string
          enum: [active, inactive]
      required:
        - name
        - script
        - trigger
        - status
    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        script:
          type: string
        trigger:
          type: string
          enum: [manual, scheduled, event]
        status:
          type: string
          enum: [active, inactive]
    AutomationRunResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        runAt:
          type: string
          format: date-time
      required:
        - success
        - output
        - runAt
...
    Habit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        frequency:
          type: string
        streak:
          type: integer
        bestStreak:
          type: integer
        totalCompletions:
          type: integer
        color:
          type: string
        icon:
          type: string
        lastCompleted:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - frequency
        - streak
        - bestStreak
        - totalCompletions
        - color
        - icon
        - createdAt
    CreateHabitBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: string
        color:
          type: string
        icon:
          type: string
      required:
        - name
    UpdateHabitBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: string
        color:
          type: string
        icon:
          type: string
...